cloud-functions/batch_orchestrator_function/src/requirements.txt
functions-framework
google-auth
requests
google-cloud-aiplatform
